--- src/BoardList.js
@@ -1,12 +1,0 @@
-import React, {  } from 'react';
-
-const BoardList = ({list}) => {
-    
-    return (
-        <div>
-            보드리스트
-        </div>
-    );
-};
-
-export default BoardList;
